requirements.txt
chardet==2.0.1
Django==1.6.2
django-bootstrap3==3.2.1
django-ckeditor-updated==4.2.3
django-extensions==1.3.3
django-filer==0.9.5
django-formset-js==0.4.0
django-haystack==2.1.0
django-jquery-js==2.0.3
django-modeltranslation==0.7.3
django-mptt==0.5.5
django-polymorphic==0.5.3
django-taggit==0.12.3
django-taggit-templatetags==0.4.5.dev0
django-templatetag-sugar==1.0
easy-thumbnails==1.5
elasticsearch==1.0.0
fpconst==0.7.2
MySQL-python==1.2.5
pbr==0.10.8
Pillow==2.4.0
pyelasticsearch==0.6.1
pygraphviz==1.2
python-apt==0.8.8.2
python-debian==0.1.21
python-debianbts==1.11
python-ldap==2.4.13
python-memcached==1.53
reportbug==6.4.4
requests==2.2.1
simplejson==3.3.3
six==1.4.1
SOAPpy==0.12.0
South==0.8.4
stevedore==1.3.0
urllib3==1.7.1
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
